Refactoring after initial commit, Added new implementation for Flight, Baggage and Cargo Added new use cases for Flight flow

+++ core/src/main/java/org/flightcargo/core/ports/BaggageRepositoryService.java
@@ -6,7 +6,7 @@ import java.util.Collection;
 
 public interface BaggageRepositoryService {
 
-    Collection<Baggage> getBaggageByFlight(int flightId);
+    Collection<Baggage> getBaggageByFlight(Long flightId);
 
     void addBaggage(Baggage baggage);
